+++ tests/testthat/test-typst.R
@@ -2,20 +2,30 @@ local_edition(2)
 
 test_that("to_typst basic table structure", {
   ht <- hux(a = 1:2, b = 3:4, add_colnames = FALSE)
+  valign(ht) <- NA
   res <- to_typst(ht)
   expected <- paste0(
     "#table(\n",
     "  columns: (auto, auto)\n",
     ")[\n",
-    "  cell(align: right)[1] cell(align: right)[3]\n",
-    "  cell(align: right)[2] cell(align: right)[4]\n",
+    "  cell(align: (right, top), inset: 6pt)[1] cell(align: (right, top), inset: 6pt)[3]\n",
+    "  cell(align: (right, top), inset: 6pt)[2] cell(align: (right, top), inset: 6pt)[4]\n",
     "]\n"
   )
   expect_identical(res, expected)
 })
 
+test_that("to_typst handles vertical alignment", {
+  ht <- hux(a = 1:2, b = 3:4, add_colnames = FALSE)
+  valign(ht) <- matrix(c("middle", "top", "top", "bottom"), 2, 2)
+  res <- to_typst(ht)
+  expect_match(res, "cell\\(align: \\(right, center\\), inset: 6pt\\)\\[1\\]")
+  expect_match(res, "cell\\(align: \\(right, bottom\\), inset: 6pt\\)\\[4\\]")
+})
+
 test_that("to_typst maps properties", {
   ht <- hux(a = 1:3, b = 4:6, c = 7:9, add_colnames = FALSE)
+  valign(ht) <- NA
   caption(ht) <- "A cap"
   col_width(ht) <- c(.2, .3, .5)
   width(ht) <- 0.5
@@ -30,6 +40,11 @@ test_that("to_typst maps properties", {
   font_size(ht)[1, 1] <- 12
   font(ht)[1, 1] <- "Courier"
 
+  left_padding(ht)[1, 3] <- 1
+  right_padding(ht)[1, 3] <- 2
+  top_padding(ht)[1, 3] <- 3
+  bottom_padding(ht)[1, 3] <- 4
+
   res <- to_typst(ht)
 
   expect_match(res, "caption: \\[A cap\\]")
@@ -38,10 +53,11 @@ test_that("to_typst maps properties", {
   expect_match(res, "height: 25\\.000%")
   expect_match(res, "colspan: 2")
   expect_match(res, "rowspan: 2")
-  expect_match(res, "align: right")
+  expect_match(res, "align: \\(right, top\\)")
   expect_match(res, "fill: rgb")
   expect_match(res, "stroke: \\(top: 1pt \\+ solid \\+ rgb")
   expect_match(res, "text\\(weight: \"bold\", style: \"italic\", size: 12pt, family: \"Courier\"\\)\\[1\\]")
+  expect_match(res, "inset: \\(top: 3pt, right: 2pt, bottom: 4pt, left: 1pt\\)")
 })
 
 test_that("print_typst outputs to stdout", {